requirements.txt
kagglehub
pandas
numpy
langchain
langchain-openai
langchain-groq
python-dotenv
python-docx
